HBASE-969  Won't when storefile > 2G.
git-svn-id: https://svn.apache.org/repos/asf/hadoop/hbase/trunk@709519 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/hadoop/hbase/io/BlockFSInputStream.java
@@ -151,7 +151,7 @@ public class BlockFSInputStream extends FSInputStream {
   }
 
   private synchronized void blockSeekTo(long target) throws IOException {
-    int targetBlock = (int) (target / blockSize);
+    long targetBlock = target/blockSize;
     long targetBlockStart = targetBlock * blockSize;
     long targetBlockEnd = Math.min(targetBlockStart + blockSize, fileLength) - 1;
     long blockLength = targetBlockEnd - targetBlockStart + 1;
